Admin implementado

+++ database/migrations/2026_07_28_133309_create_categorias_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nome')->unique();
             $table->text('descricao')->nullable();
             $table->boolean('ativo')->default(true);
-            $table->unsignedBigInteger('ordem_exibicao')->default(0);
+            $table->unsignedInteger('ordem_exibicao')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
